backend/requirement.txt
Flask==2.0.3
requests==2.26.0
pymongo==3.12.1
google-api-python-client==2.25.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
python-dotenv==0.19.0
